Move out handleSettings, chart, help into header hook
