+++ src/atoms/headerAtoms.ts
@@ -1,0 +1,8 @@
+import { atom } from 'jotai';
+
+
+export const chartsOpenAtom = atom<boolean>(false);
+
+export const helpOpenAtom = atom<boolean>(false);
+
+export const settingsOpenAtom = atom<boolean>(false);
