+++ src/components/Header/ChartsPopup.tsx
@@ -6,17 +6,18 @@ import {AppContext} from "../../contexts/AppContext";
 import grey from "@mui/material/colors/grey";
 import { useAtom } from "jotai";
 import { displayWordsAtom } from "../../atoms/bibleDisplayAtoms";
+import { chartsOpenAtom } from "../../atoms/headerAtoms";
 
 const ChartsPopup = () => {
   const [displayWords] = useAtom(displayWordsAtom)
+  const [chartsOpen, setChartsOpen] = useAtom(chartsOpenAtom);
 
   const theme = useTheme();
   const context = useContext(AppContext);
   if (!context) {
     return null;
   }
-  const {
-    chartsOpen, setChartsOpen, testWordIndices, currentIndex, selectedTesters
+  const { testWordIndices, currentIndex, selectedTesters
   } = context;
